ci: enable easy way to skip acceptance test
If we add the tag "skip" in the BUILD.bazel file
of an acceptance test, this will be skipped in ci.

Also
- unzips the outputs.zip from the bazel folder in order
to be easier reachable once we download the logs.

For the records, a solution based on the following

```
--test_tag_filters=<comma-separated list of options> default: ""

Specifies a comma-separated list of test tags. Each tag can be
optionally preceded with '-' to specify excluded tags. Only those test
targets will be found that contain at least one included tag and do not
contain any excluded tags. This option affects --build_tests_only
behavior and the test command.
```
which would lead to less bash was not possible because bazel test
if no test exists returns with return code 4.