gateway-testing: add fake configuration spec
This adds the configuration specification for the fake gateway runner.
The idea of the fake gateway runner is to provide a way to configure a
gateway dataplane without a running gateway control plane. This can be
helpful for dataplane only tests, like performance tests.

The JSON schema was generated from the JSON using
https://extendsclass.com/json-schema-validator.html after generation the
description fields were added to the schema.

GitOrigin-RevId: 0b91740e661e0c002f344516f4b641e7d5892dcc