posix-router: fix bug for SCMP messages on cross-over paths
This PR fixes a bug where the posix-router wrongly updates the segment
ID when generating SCMP messages.

This bug only appears for packets that cross over from one path segment
to the next in the AS of the router, and the packet is received from
an other AS internal router. For cross over paths with the same ingress
and egress router, this bug does not appear.

Preparing the SCMP packet is streamlined and we operate directly on the current path
and SCION layer.

GitOrigin-RevId: df0700b8b4de28cb129a594aa2e26be13bd20803