gateway: dynamically announce routes in the main routing table
The additional routing table used by the gateway makes configuration
very awkward. Additionally, due to the firewall marking/rules, some of the
benefits of dynamic routing are lost.

With this PR, the additional gateway routing table is dropped, and the
learned networks are announced dynamically on the main routing table.

GitOrigin-RevId: fe4bf46b9b845eee12fde2edb31bb72011b4be5d