oai: add `/segments/{segment-id}` and `/segments/{segment-id}/blob` endpoint
Adds the `/segments/{segment-id}` and `/segments/{segment-id}/blob` endpoints.
It is possible to request multiple segments with a comma separated list.

The result for the `/segments/{segment-id}` contains a list of segment descriptions with the following fields:
- id: the segment ID.
- hops: list of (interface, isd_as) pairs.
- expiration: the minimum over all hop expiration times
- timestamp: the segment info timestamp
- last_updated: the last time this segment was updated in the database.

The result of the `/segments/{segment-id}/blob` endpoint is the raw protobuf segments encoded in PEM.

GitOrigin-RevId: ad075728f550eae302536696e116029917039f28