trust: reduce DB interface
Reduce the DB interface of the trust DB to the necessary methods.
The `Closer` and `LimitSetter` methods are outside of the concerns of
the trust package.

This will also pave the way to extend the trust DB with additional
methods for the API which should not be visible by the trust package.

GitOrigin-RevId: 4babdb5f49a9eefea958c4c0c6139b9e5771e898