ca: add support for CMS enveloped chain renewal requests
This PR adds support for CMS enveloped chain renewal requests. For the moment, both formats, i.e., chain renewal enveloped in a protobuf signed message and in a CMS message are supported, with precedence given to the CMS envelop.
Also, reworked crypto generation for the new tests and regenerated the legacy crypto.

GitOrigin-RevId: d59fc67299f8345d4937e50aba08d77c127b2f6b