gateway: use tunnel.src_ipv4/6 for route src hint
Use the existing but previously ignored `tunnel.src_ipv4` / `tunnel.src_ipv6` configuration options as the src IP hint in the IP routes. Explicitly pass different source address for IPv4/6 to `publishingRoutingTable` and use the appropriate address based on the IP version of the specific prefix.

Also, fix naming of "NextHop" member in routemgr.Route -- this is the source address hint for the route.

Finally, fix a small typo in the manual.

Fixes #3980
Closes #3983

GitOrigin-RevId: e88582030828cfcb1c14329fab54299104dcab48