cppki: change semantics of ExtractIA
Previously, `cppki.ExtractIA` would extract the ISD-AS from the
distinguished name if it existed. Otherwise, nil was returned.
This API is not straight-forward and prone to bugs in client code.

It turns out that in almost all cases, we want to return an error if
no ISD-AS is present. The cases where we want the old semantics with a
nil return is restricted to `cppki` internal client code.

Thus, `cppki.ExtractIA` now returns an error if there is no ISD-AS in the
distinguished name, or if it fails to parse.

GitOrigin-RevId: 0b33bf07b8c7a705023ab3ab5532894b189725b4

Original commit amended for DRKey.
Co-authored-by: Matthias Frei <matthias.frei@inf.ethz.ch>