cs: limit retries for seg requests
This fixes a regression introduced in #3993; certain types of errors, in
particular `ErrNotReachable` indicating that there is no path to the
server where the request should be made at, occur very quickly and will
not be fixed by retrying.
When each try fails very quickly, the retry mechanism would just busy
wait until the request context ends.

Abort immediately in case of `ErrNotReachable`.
Also limit the number of retries to catch any other cases of quickly
failing attempts; the number of allowed retries is currently 20, which
should be ample to keep exploring paths even when the request deadline
is generous, but low enough to make sure that run away retries are not
too much of a nuisance.

The retry limit applied for seg reqs from sciond to the CS is set to 1;
these retries are not useful here (TCP, no paths to explore).

Closes #4000

GitOrigin-RevId: 06867e4402dba27439b285fdad21c290fe9fad65