ping: attach error details if there are interfaces down on the path
When processing the response to an echo request, create specialized errors for replies that indicate that interfaces were down on the path.

The following error detail is added:
- SCMPExternalInterfaceDown: the offending ISD-AS and the interface identifier
- SCMPInternalConnectivityDown: the offending ISD-AS and the ingress/egress interface identifiers

This helps pinpoint connectivity issues on a SCION path that is used for pinging.

An example output for the `scion` tool after the change looks like this:

    ERROR: external interface is down isd_as="1-ff00:0:110" interface="42"

Closes #4011

GitOrigin-RevId: 5c8f9bb4354610c2821001f3fbb477e012ceb5b6