scion-pki: make 'certificate renew' command easier to use
Simplify the `scion-pki certificate renew` command and make it easier
to use. With this refactor this command can be used as a periodic
cronjob directly.

This change breaks previous usages of the `scion-pki certificate renew` command.

This change bumps the scion-pki version to v0.5.0.

```
'renew' requests a renewed AS certificate from a remote CA control service.

The provided <chain-file> and <key-file> are used to sign the CSR. They must be
valid and verifiable by the CA in order for the request to be served.

The renewed certificate chain is requested with a fresh private key, unless the
--reuse-key flag is set.

The TRCs are used to validate and verify the renewed certificate chain. If the
chain is not verifiable with any of the active TRCs, the certificate chain and,
if applicable, the fresh private key are written to the provided file paths with
the '.unverified' suffix.

The resulting certificate chain is written to the file system, either to
<chain-file> or to --out, if specified.

The fresh private key is is written to the file stystem, either to <key-file>
or to --out-key, if specified.

Files are not allowed to be overwritten, by default. Either you have to specify
the --out and --out-key flags explicitely, or specify the --force or --backup
flags. In case the --backup flag is set, every file that would be overwritten is
renamed to contain a local execution time timestamp before the file extension.
E.g., <filename-base>.<YYYY-MM-DD-HH-MM-SS>.<filename-ext>.

This command supports the --expires-in flag in order for it to be run in a
periodic task runner (e.g., cronjob). The flag indicates the acceptable remaining
time before certificate expiration. If the remaining time is larger or equal to
the specified value, the command immediately exits with code zero. If the
remaining time is less than the specified value, a renewal run is executed.
The time can either be specified as a time duration or a relative factor of the
existing certificate chain. For the time duration, the following units are
supported: d, h, m, s. The relative factor is supplied as a floating point
number. For example, a factor of 0.75 indicates that the certificate chain
should be renewed after one quarter of its lifetime has passed, and it still
has three quarters of its validity period until it expires.

Unless a subject template is specified, the subject of the existing certificate
chain is used as the subject for the renewal request.

The template is expressed in JSON. A valid example:

  {
    "common_name": "1-ff00:0:110 AS certificate",
    "country": "CH",
    "isd_as": "1-ff00:0:110"
  }

All configurable fields with their type are defined by the following JSON
schema. For more information on JSON schemas, see https://json-schema.org/.

  {
    "type": "object",
    "properties": {
      "isd_as":              { "type": "string" },
      "common_name":         { "type": "string" },
      "country":             { "type": "string" },
      "locality":            { "type": "string" },
      "organization":        { "type": "string" },
      "organizational_unit": { "type": "string" },
      "postal_code":         { "type": "string" },
      "province":            { "type": "string" },
      "serial_number":       { "type": "string" },
      "street_address":      { "type": "string" },
    },
    "required": ["isd_as"]
  }

```

GitOrigin-RevId: 81df319e0960462a801bcfbb841e3a0df03ad997