acceptance: avoid making the generated topology an action
Actions in bazel are supposed to be deterministic. They are cached
indefinitely. From the bazel documentation:

> Actions are comparable to pure functions: They should depend only on
> the provided inputs, and avoid accessing computer information, username,
> clock, network, or I/O devices (except for reading inputs and writing
> outputs). This is important because the output will be cached and reused.

The topologies we create contain crypto material. With recent changes,
this material is only valid for 3 days. As a result, tests can start
randomly failing if the topology artifact is still cached and the 3 days
have passed. This is for example possible in CI, if topogen and scion-pki
have not been touched in 3 days.

With this changes, the topology generation is moved to the acceptance
test setup logic.

GitOrigin-RevId: dff0f546db49e7965d088b6bb424dcda189933a4